+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the React Track Player Web project will be documented in this file.
 
+## 1.0.2 - 2025-04-21
+
+### Added
+
+- 📚 Published official documentation site at https://itslhuis.github.io/react-track-player-web/
+
+### Fixed
+
+- 🏠 Fixed the `homepage` field in `package.json` to correctly point to the documentation site
+
 ## 1.0.1 - 2025-04-20
 
 ### Added